requirements.txt
Flask==2.3.2
cohere==4.5.1
python-dotenv==1.0.1